♻️ refactor : exception 패키지 common 패키지로 이동 및 board,comment responseDto 병합
