+++ src/main/java/com/prac/music/common/exception/BoardRuntimeException.java
@@ -1,4 +1,4 @@
-package com.prac.music.exception;
+package com.prac.music.common.exception;
 
 public class BoardRuntimeException extends RuntimeException {
 	public BoardRuntimeException(String message) {
