+++ src/main/java/com/prac/music/common/exception/CommentRuntimeException.java
@@ -1,4 +1,4 @@
-package com.prac.music.exception;
+package com.prac.music.common.exception;
 
 public class CommentRuntimeException extends RuntimeException {
 	public CommentRuntimeException(String message) {
