+++ src/main/java/com/prac/music/common/exception/UnauthorizedAccessException.java
@@ -1,4 +1,4 @@
-package com.prac.music.exception;
+package com.prac.music.common.exception;
 
 public class UnauthorizedAccessException extends RuntimeException {
 	public UnauthorizedAccessException(String message) {
