+++ src/main/java/com/prac/music/common/exception/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.prac.music.exception;
+package com.prac.music.common.exception;
 
 public class UserNotFoundException extends RuntimeException {
     public UserNotFoundException(String message) {
